+++ src/main/kotlin/com/github/bgomar/consolelogger/UpdateNotify.kt
@@ -5,13 +5,14 @@ import com.intellij.ide.plugins.PluginManagerCore
 import com.intellij.notification.NotificationType
 import com.intellij.openapi.extensions.PluginId
 import com.intellij.openapi.project.Project
-import com.intellij.openapi.startup.StartupActivity
+import com.intellij.openapi.startup.ProjectActivity
 
-class UpdateNotify : StartupActivity {
+class UpdateNotify : ProjectActivity {
+
   private val plugin = PluginManagerCore.getPlugin(PluginId.getId("com.github.bgomar.consolelogger"))!!
 
-  override fun runActivity(project: Project) {
-    val settings = ConsoleLoggerSettings.getInstance();
+  override suspend fun execute(project: Project) {
+    val settings = ConsoleLoggerSettings.getInstance()
     if (settings.version == "Unknown") {
       settings.version = plugin.version
       showUpdate(project)
@@ -21,23 +22,22 @@ class UpdateNotify : StartupActivity {
     }
   }
 
-  private val updateContent: String by lazy {
-    //language=HTML
-    """
-    🐛 Bugfix for Saving 😁👌<br><br>
-    Sorry it took so long, it was quite a hard task.<br>
-    🤏 Small changes to defaults: <br> 
-    - changed some of the input field sizes
-    - added option for second preset of loggers
-    """
-  }
-
   private fun showUpdate(project: Project) {
     val notification = createNotification(
       "ConsoleLogger plugin updated to version ${plugin.version}",
-      updateContent,
+      """
+            🐛 Bugfix for Saving 😁👌<br><br>
+            Sorry it took so long, it was quite a hard task.<br>
+            🤏 Small changes to defaults: <br>
+            - Changed input field sizes<br>
+            - Added second logger preset option<br>
+            - Fixed Issue with Code Insertion Position and <br>
+              Log Line Number Updates in Plugin<br>
+            - Added the Re-check line numbers function<br>
+            """,
       NotificationType.INFORMATION
     )
     showFullNotification(project, notification)
   }
 }
+